# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Planner API",
        "description": "Event Planner API is a Laravel API designed to manage events. It allows users to create, read, update, and delete events. Additionally, the API provides authentication features, including email verification and Google sign-up.",
        "contact": {
            "name": "Taras Maltsev",
            "url": "https://github.com/TarasIT",
            "email": "taras.maltsev@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://event-planner-api.onrender.com/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/auth/google/redirect": {
            "get": {
                "tags": [
                    "Google Authentication"
                ],
                "summary": "Redirect to Google for authentication",
                "description": "Redirects the user to Google for authentication",
                "operationId": "aa5527a0ef4a252c1a318102ec441169",
                "responses": {
                    "302": {
                        "description": "Redirect to Google"
                    },
                    "500": {
                        "description": "Failed google redirect. Please try later."
                    }
                }
            }
        },
        "/auth/google/callback": {
            "get": {
                "tags": [
                    "Google Authentication"
                ],
                "summary": "Handle Google callback",
                "description": "Handles the callback from Google after authentication",
                "operationId": "14def10fcc3c6f92fbc8348c8e977a08",
                "responses": {
                    "200": {
                        "description": "Successfully handled Google callback"
                    },
                    "500": {
                        "description": "Failed google callback. Please try later."
                    }
                }
            }
        },
...
        "/events/{id}": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "Gets the event by id",
                "operationId": "d9310ef60622d67beca4fb63269132b8",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ResponseEventScheme"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Event with id={id} is not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve an event",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to retrieve an event. Please, try later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Event"
                ],
                "summary": "Updates the event by id",
                "description": "Updates the event by id",
                "operationId": "5ac2e66032000d5eea414ddc43c2fbc5",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEvent"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ResponseEventScheme"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Event with id={id} is not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to update an event",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to update an event. Please, try later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Event"
                ],
                "summary": "Deletes the event by id",
                "operationId": "66c27749b47fbda989fe29b76bab4a24",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Event with id={id} is not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to delete an event",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to delete an event. Please, try later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdateEvent": {
                "required": [
                    "title",
                    "date",
                    "time"
                ],
                "properties": {
                    "title": {
                        "description": "Title of the event",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the event",
                        "type": "string"
                    },
                    "date": {
                        "description": "Date of the event",
                        "type": "string"
                    },
                    "time": {
                        "description": "Time of the event",
                        "type": "string"
                    },
                    "location": {
                        "description": "Location of the event",
                        "type": "string"
                    },
                    "category": {
                        "description": "Category of the event",
                        "type": "string"
                    },
                    "priority": {
                        "description": "Priority of the event",
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ]
                    },
                    "picture": {
                        "description": "Base64 encoded string of the event picture",
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "title": "Meeting",
                    "description": "This meeting is very important",
                    "date": "01/07/2024",
                    "time": "09:00 AM",
                    "location": "Kyiv",
                    "category": "Business",
                    "picture": "base64-encoded-string",
                    "priority": "high"
                }
            },
            "ResponseEventScheme": {
                "properties": {
                    "id": {
                        "description": "Event ID",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Title of the event",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the event",
                        "type": "string"
                    },
                    "date": {
                        "description": "Date of the event",
                        "type": "string"
                    },
                    "time": {
                        "description": "Time of the event",
                        "type": "string"
                    },
                    "location": {
                        "description": "Location of the event",
                        "type": "string"
                    },
                    "category": {
                        "description": "Category of the event",
                        "type": "string"
                    },
                    "priority": {
                        "description": "Priority of the event",
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ]
                    },
                    "picture": {
                        "description": "Picture of the event",
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "id": "4",
                    "title": "Meeting",
                    "description": "This meeting is very important",
                    "date": "01/07/2024",
                    "time": "09:00 AM",
                    "location": "Kyiv",
                    "category": "Business",
                    "picture": "path/to/the/picture/in/the/cloud/storage",
                    "priority": "high"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}